requirements.txt
pillow~=9.4.0
fonttools~=4.25.0
future~=0.18.3
beautifulsoup4~=4.12.2